refactor(cli): move cmd/agent under cmd/internal to mirror command tree
