+++ cmd/helper/helper.go
@@ -1,11 +1,11 @@
 package helper
 
 import (
-	"github.com/devsy-org/devsy/cmd/agent"
 	"github.com/devsy-org/devsy/cmd/flags"
 	"github.com/devsy-org/devsy/cmd/helper/http"
 	"github.com/devsy-org/devsy/cmd/helper/json"
 	"github.com/devsy-org/devsy/cmd/helper/strings"
+	"github.com/devsy-org/devsy/cmd/internal/agent"
 	"github.com/spf13/cobra"
 )
