+++ cmd/internal/agent/container/deferred_hooks.go
