+++ cmd/internal/agent/container/deferred_hooks_windows.go
