+++ cmd/internal/agent/container/post_attach_windows.go
