+++ cmd/internal/agent/container/setup_windows.go
