+++ cmd/internal/agent/container/vscode_async.go
