+++ cmd/internal/agent/docker_credentials.go
