+++ cmd/internal/agent/git_ssh_signature.go
