+++ cmd/internal/agent/git_ssh_signature_helper.go
