+++ cmd/internal/agent/workspace/delete.go
