+++ cmd/internal/agent/workspace/delete_test.go
