+++ cmd/internal/agent/workspace/docker_darwin_test.go
