+++ cmd/internal/agent/workspace/install_dotfiles.go
