+++ cmd/internal/agent/workspace/setup_gpg.go
