+++ cmd/internal/agent/workspace/status.go
