+++ cmd/internal/agent/workspace/stop.go
