+++ cmd/internal/agent/workspace/update_config.go
