+++ cmd/internal/internal.go
@@ -1,9 +1,9 @@
 package cmdinternal
 
 import (
-	"github.com/devsy-org/devsy/cmd/agent"
 	"github.com/devsy-org/devsy/cmd/flags"
 	"github.com/devsy-org/devsy/cmd/helper"
+	"github.com/devsy-org/devsy/cmd/internal/agent"
 	"github.com/spf13/cobra"
 )
