+++ cmd/pro/provider/provider.go
@@ -3,7 +3,7 @@ package provider
 import (
 	"os"
 
-	"github.com/devsy-org/devsy/cmd/agent"
+	"github.com/devsy-org/devsy/cmd/internal/agent"
 	"github.com/devsy-org/devsy/cmd/pro/flags"
 	"github.com/devsy-org/devsy/cmd/pro/provider/create"
 	"github.com/devsy-org/devsy/cmd/pro/provider/get"
